added fixture data management command to crm app, removed migrations from repo, added project reset shell script, updated .gitignore, and fixed some issues with the models
