+++ happyteams/happyteams/settings.py
@@ -122,3 +122,7 @@ USE_TZ = True
 # https://docs.djangoproject.com/en/1.11/howto/static-files/
 
 STATIC_URL = '/static/'
+
+
+FIXTURE_SUPER_USERNAME = 'admin'
+FIXTURE_SUPER_PASSWORD = 'admin'
